Add NetworkParametersStorageInternal
