+++ core/src/main/kotlin/net/corda/core/node/services/NetworkParametersStorage.kt
@@ -1,5 +1,6 @@
 package net.corda.core.node.services
 
+import net.corda.core.CordaInternal
 import net.corda.core.DoNotImplement
 import net.corda.core.crypto.SecureHash
 import net.corda.core.internal.SignedDataWithCert
@@ -26,17 +27,4 @@ interface NetworkParametersStorage {
      * get them from network map.
      */
     fun readParametersFromHash(hash: SecureHash): NetworkParameters?
-
-    /**
-     * Return parameters epoch for the given parameters hash. Null if there are no parameters for this hash in the storage and we are unable to
-     * get them from network map.
-     */
-    fun getEpochFromHash(hash: SecureHash): Int?
-
-    /**
-     * Save signed network parameters data. Internally network parameters bytes should be stored with the signature.
-     * It's because of ability of older nodes to function in network where parameters were extended with new fields.
-     * Hash should always be calculated over the serialized bytes.
-     */
-    fun saveParameters(signedNetworkParameters: SignedDataWithCert<NetworkParameters>)
 }
