+++ node/src/main/kotlin/net/corda/node/internal/DBNetworkParametersStorage.kt
@@ -24,13 +24,28 @@ import org.apache.commons.lang.ArrayUtils
 import java.security.cert.X509Certificate
 import javax.persistence.*
 
+interface NetworkParametersStorageInternal : NetworkParametersStorage {
+    /**
+     * Return parameters epoch for the given parameters hash. Null if there are no parameters for this hash in the storage and we are unable to
+     * get them from network map.
+     */
+    fun getEpochFromHash(hash: SecureHash): Int?
+
+    /**
+     * Save signed network parameters data. Internally network parameters bytes should be stored with the signature.
+     * It's because of ability of older nodes to function in network where parameters were extended with new fields.
+     * Hash should always be calculated over the serialized bytes.
+     */
+    fun saveParameters(signedNetworkParameters: SignedDataWithCert<NetworkParameters>)
+}
+
 class DBNetworkParametersStorage(
         cacheFactory: NamedCacheFactory,
         private val database: CordaPersistence,
         // TODO It's very inefficient solution (at least at the beginning when node joins without historical data)
         // We could have historic parameters endpoint or always add parameters as an attachment to the transaction.
         private val networkMapClient: NetworkMapClient?
-) : NetworkParametersStorage, SingletonSerializeAsToken() {
+) : NetworkParametersStorageInternal, SingletonSerializeAsToken() {
     private lateinit var trustRoot: X509Certificate
 
     companion object {
