+++ node/src/main/kotlin/net/corda/node/services/network/NetworkMapUpdater.kt
@@ -9,10 +9,10 @@ import net.corda.core.messaging.DataFeed
 import net.corda.core.messaging.ParametersUpdateInfo
 import net.corda.core.node.NetworkParameters
 import net.corda.core.node.services.KeyManagementService
-import net.corda.core.node.services.NetworkParametersStorage
 import net.corda.core.serialization.serialize
 import net.corda.core.utilities.contextLogger
 import net.corda.core.utilities.minutes
+import net.corda.node.internal.NetworkParametersStorageInternal
 import net.corda.node.services.api.NetworkMapCacheInternal
 import net.corda.node.services.config.NetworkParameterAcceptanceSettings
 import net.corda.node.utilities.NamedThreadFactory
@@ -35,7 +35,7 @@ class NetworkMapUpdater(private val networkMapCache: NetworkMapCacheInternal,
                         private val networkMapClient: NetworkMapClient?,
                         private val baseDirectory: Path,
                         private val extraNetworkMapKeys: List<UUID>,
-                        private val networkParametersStorage: NetworkParametersStorage
+                        private val networkParametersStorage: NetworkParametersStorageInternal
 ) : AutoCloseable {
     companion object {
         private val logger = contextLogger()
